refactor git operations to module
